Working account query with flags and thresholds

+++ src/account/schema/account.type.ts
@@ -1,10 +1,11 @@
 import { gql } from "apollo-server";
 
+// NOTE: GraphQL treats INT as 32-bit numbers. sequenceNumber is 52-bit BigInt hence have to use Float.
 const accountType = gql`
   type Account {
     id: AccountID!
     balance: Float!
-    sequenceNumber: Int!
+    sequenceNumber: Float!
     numSubentries: Int!
     inflationDest: AccountID
     homeDomain: String
